requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
SQLAlchemy>=2
asyncpg
alembic
python-jose[cryptography]
python-dotenv